Refactor language handling
