+++ frontend/src/utils/languageUtil.js
@@ -1,5 +1,10 @@
 // TODO: translations for language names?
 
+const defaultLanguage = {
+  short: 'en',
+  name: 'English'
+}
+
 export const languages = [
   {
     short: 'en',
@@ -13,11 +18,6 @@ export const languages = [
 
 export const getByName = name => {
   return languages.filter(l => l.name === name)[0]
-}
-
-const defaultLanguage = {
-  short: 'en',
-  name: 'English'
 }
 
 export default defaultLanguage
